Extract PluralHelper to separate file
